Split MerklePatricia.hs into internal and external interfaces.
The external interface has only "safe" functions that receive key name
literals and convert them to hashes for lookup.  The internal
interface has the unsafe, unhashed versions as well as a function to
return all key-values, and the implementations.
